Update ReadStat sources #94 #98 #117
 * Export a default format for Stata variables. Fixes #94 #98

 * Support for tagged missing values from SAS and Stata (haven will
   need to decide what, if anything, to do with these). Addresses
   but does not fix #117

 * Many improvements to SAS catalog importer
