+++ src/readstat_spss.c
@@ -17,33 +17,33 @@ static spss_type_t spss_type_strings[] = {
     { .type = SPSS_FORMAT_TYPE_DOLLAR, .string = "DOLLAR" },
     { .type = SPSS_FORMAT_TYPE_F,      .string = "F" },
     { .type = SPSS_FORMAT_TYPE_IB,     .string = "IB" },
-    { .type = SPSS_FORMAT_TYPE_PIBHEX,  .string = "PIBHEX" },
-    { .type = SPSS_FORMAT_TYPE_P,  .string = "P" },
-    { .type = SPSS_FORMAT_TYPE_PIB,  .string = "PIB" },
-    { .type = SPSS_FORMAT_TYPE_PK,  .string = "PK" },
-    { .type = SPSS_FORMAT_TYPE_RB,  .string = "RB" },
+    { .type = SPSS_FORMAT_TYPE_PIBHEX, .string = "PIBHEX" },
+    { .type = SPSS_FORMAT_TYPE_P,      .string = "P" },
+    { .type = SPSS_FORMAT_TYPE_PIB,    .string = "PIB" },
+    { .type = SPSS_FORMAT_TYPE_PK,     .string = "PK" },
+    { .type = SPSS_FORMAT_TYPE_RB,     .string = "RB" },
     { .type = SPSS_FORMAT_TYPE_RBHEX,  .string = "RBHEX" },
-    { .type = SPSS_FORMAT_TYPE_Z,  .string = "Z" },
-    { .type = SPSS_FORMAT_TYPE_N,  .string = "N" },
-    { .type = SPSS_FORMAT_TYPE_E,  .string = "E" },
-    { .type = SPSS_FORMAT_TYPE_DATE,  .string = "DATE" },
-    { .type = SPSS_FORMAT_TYPE_TIME,  .string = "TIME" },
-    { .type = SPSS_FORMAT_TYPE_DATETIME,  .string = "DATETIME" },
+    { .type = SPSS_FORMAT_TYPE_Z,      .string = "Z" },
+    { .type = SPSS_FORMAT_TYPE_N,      .string = "N" },
+    { .type = SPSS_FORMAT_TYPE_E,      .string = "E" },
+    { .type = SPSS_FORMAT_TYPE_DATE,   .string = "DATE" },
+    { .type = SPSS_FORMAT_TYPE_TIME,   .string = "TIME" },
+    { .type = SPSS_FORMAT_TYPE_DATETIME,.string = "DATETIME" },
     { .type = SPSS_FORMAT_TYPE_ADATE,  .string = "ADATE" },
     { .type = SPSS_FORMAT_TYPE_JDATE,  .string = "JDATE" },
     { .type = SPSS_FORMAT_TYPE_DTIME,  .string = "DTIME" },
     { .type = SPSS_FORMAT_TYPE_WKDAY,  .string = "WKDAY" },
     { .type = SPSS_FORMAT_TYPE_MONTH,  .string = "MONTH" },
-    { .type = SPSS_FORMAT_TYPE_MOYR,  .string = "MOYR" },
-    { .type = SPSS_FORMAT_TYPE_QYR,  .string = "QYR" },
-    { .type = SPSS_FORMAT_TYPE_WKYR,  .string = "WKYR" },
-    { .type = SPSS_FORMAT_TYPE_PCT,  .string = "PCT" },
-    { .type = SPSS_FORMAT_TYPE_DOT,  .string = "DOT" },
-    { .type = SPSS_FORMAT_TYPE_CCA,  .string = "CCA" },
-    { .type = SPSS_FORMAT_TYPE_CCB,  .string = "CCB" },
-    { .type = SPSS_FORMAT_TYPE_CCC,  .string = "CCC" },
-    { .type = SPSS_FORMAT_TYPE_CCD,  .string = "CCD" },
-    { .type = SPSS_FORMAT_TYPE_CCE,  .string = "CCE" },
+    { .type = SPSS_FORMAT_TYPE_MOYR,   .string = "MOYR" },
+    { .type = SPSS_FORMAT_TYPE_QYR,    .string = "QYR" },
+    { .type = SPSS_FORMAT_TYPE_WKYR,   .string = "WKYR" },
+    { .type = SPSS_FORMAT_TYPE_PCT,    .string = "PCT" },
+    { .type = SPSS_FORMAT_TYPE_DOT,    .string = "DOT" },
+    { .type = SPSS_FORMAT_TYPE_CCA,    .string = "CCA" },
+    { .type = SPSS_FORMAT_TYPE_CCB,    .string = "CCB" },
+    { .type = SPSS_FORMAT_TYPE_CCC,    .string = "CCC" },
+    { .type = SPSS_FORMAT_TYPE_CCD,    .string = "CCD" },
+    { .type = SPSS_FORMAT_TYPE_CCE,    .string = "CCE" },
     { .type = SPSS_FORMAT_TYPE_EDATE,  .string = "EDATE" },
     { .type = SPSS_FORMAT_TYPE_SDATE,  .string = "SDATE" }
 };
